Organize into common lib importable from anywhere
