--- math_utils/__init__.py
@@ -1,4 +1,0 @@
-__all__ = []
-
-import mobject_geometry
-__all__.extend(mobject_geometry.__all__)
