+++ setup.py
@@ -1,0 +1,8 @@
+# kudos to https://stackoverflow.com/a/50193944/10717280
+
+from setuptools import setup, find_packages
+
+
+setup(name='rc_lib', version='0.1', packages=[
+    "rc_lib"
+])
